requirements.txt
pyNeko>=3.0
pyyaml>=3.12
xlwt>=1.2.0